LPD-99982 Migrate the export report entries UI tests to the revamp
